chore: move config behind checks (#1211)
