+++ packages/backend/src/graphql/query-resolvers.ts
@@ -3,7 +3,7 @@ import getApps from './queries/get-apps';
 import getAutomatischInfo from './queries/get-automatisch-info';
 import getBillingAndUsage from './queries/get-billing-and-usage.ee';
 import getConnectedApps from './queries/get-connected-apps';
-import getConfig from './queries/get-config';
+import getConfig from './queries/get-config.ee';
 import getCurrentUser from './queries/get-current-user';
 import getDynamicData from './queries/get-dynamic-data';
 import getDynamicFields from './queries/get-dynamic-fields';
