+++ packages/web/src/graphql/queries/get-config.ee.ts
